finished core data migration; added backgrounding queue for migration and download
known bug: My Trips table view (SaveTripsViewController) still having
cell-rendering problems on tap: cell will render with content of an
additional cell. reuse is likely the culpret.

now using actual device hashes for test distribution.

handle core data migration so that it:
-handles interruptions better. if the app is backgrounded, it should
finish in the task in the background
-if the task doesn't finish, it will remove the partially migrated new
store and start over.

improved the download trips process to finish adding a given trip if
the app is backgrounded. if the job doesn't finish in that time-frame,
user will need to re-tap 'download' to fetch remaining trips.