Added a comfort picker and changed server code
Removed two trip purpose options as per request.  Added a comfort
uipickerview to the details page that allows the user to rate their
trip comfort 1-5 with corresponding text and icons.  Changed the code
that posts the data the server to correspond with our php code and
database (that now includes comfort).